+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.dto;
+package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
