+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,8 +2,9 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.dto.ReservationTimeRequest;
-import roomescape.domain.dto.ReservationTimeResponse;
+import roomescape.domain.ReservationTimeDao;
+import roomescape.dto.ReservationTimeRequest;
+import roomescape.dto.ReservationTimeResponse;
 import roomescape.domain.ReservationTime;
 
 @Service
