+++ src/test/java/roomescape/domain/dao/JdbcReservationTimeDaoTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.domain.ReservationTime;
-import roomescape.service.ReservationTimeDao;
+import roomescape.domain.ReservationTimeDao;
 import roomescape.service.Fixtures;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
