+++ src/test/java/roomescape/service/FakeReservationDao.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import roomescape.domain.Reservation;
+import roomescape.domain.ReservationDao;
 
 public class FakeReservationDao implements ReservationDao {
