+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -7,8 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.dto.ReservationTimeRequest;
-import roomescape.domain.dto.ReservationTimeResponse;
+import roomescape.domain.ReservationTimeDao;
+import roomescape.dto.ReservationTimeRequest;
+import roomescape.dto.ReservationTimeResponse;
 
 class ReservationTimeServiceTest {
